requirements.txt
nose
pyopenssl
pycrypto
pyasn1
twisted
autobahn